# apps/panel/api/openapi.yaml
openapi: 3.1.0
info:
  title: One Proxy Control Plane API
  version: 0.3.0
components:
  securitySchemes:
    controlPlaneAccessToken:
      type: apiKey
      in: header
      name: X-One-Proxy-Access-Token
    nodeToken:
      type: apiKey
      in: header
      name: X-One-Proxy-Node-Token
  parameters:
    TenantIDHeader:
      name: X-One-Proxy-Tenant-ID
      in: header
      required: true
      schema:
        type: string
      description: Required for non-super_admin tenant-scoped console requests.
...
paths:
...
  /api/proxy/route-groups:
    get:
      security:
        - controlPlaneAccessToken: []
      parameters:
        - $ref: "#/components/parameters/TenantIDHeader"
      responses:
        "200":
          description: Envelope route rule group list response
    post:
      security:
        - controlPlaneAccessToken: []
      parameters:
        - $ref: "#/components/parameters/TenantIDHeader"
      responses:
        "201":
          description: Envelope route rule group created response
...
  /api/proxy/routes:
    get:
      security:
        - controlPlaneAccessToken: []
      parameters:
        - $ref: "#/components/parameters/TenantIDHeader"
      responses:
        "200":
          description: Envelope route rule list response
    post:
      security:
        - controlPlaneAccessToken: []
      parameters:
        - $ref: "#/components/parameters/TenantIDHeader"
      responses:
        "201":
          description: Envelope route rule created response
  /api/proxy/routes/{ruleId}:
    patch:
      security:
        - controlPlaneAccessToken: []
      parameters:
        - $ref: "#/components/parameters/TenantIDHeader"
      responses:
        "200":
          description: Envelope route rule updated response
    delete:
      security:
        - controlPlaneAccessToken: []
      parameters:
        - $ref: "#/components/parameters/TenantIDHeader"
      responses:
        "200":
          description: Envelope route rule deleted response
...